mission

+++ api/misson/mission.controllers.js
@@ -20,8 +20,17 @@ exports.addMission = async (req, res) => {
 
 exports.getMission = async (req, res) => {
     try {
-    const getAllMission = await missionService.getAllMission();
-    return res.status(200).json({ msg: 'Success!', data: getAllMission });
+        const getAllMission = await missionService.getAllMission();
+        return res.status(200).json({ msg: 'Success!', data: getAllMission });
+    } catch (error) {
+        return res.status(400).json({ message: error });
+    }
+};
+
+exports.getDetailMission = async (req, res) => {
+    try {
+        const getDetailMission = await missionService.getMissionById(req.params.id);
+        return res.status(200).json({ msg: 'Success!', data: getDetailMission });
     } catch (error) {
         return res.status(400).json({ message: error });
     }
@@ -32,24 +41,28 @@ exports.updateMission = async (req, res) => {
     try {
         const getMissionById = await missionService.getMissionById(req.params.id);
         const updateMission = await missionService.updateMission(getMissionById.id, name, unit_id, description, quantity, kpiValue, startTime, endTime, manday);
-        const findResponsibleMissionDepartment = await missionService.getDepartmentMissionById(getMissionById.id, true);
-        const findResponsibleDepartment = await missionService.getDepartmentById(responsibleDepartment_id);
-        if (findResponsibleMissionDepartment) {
-            const deleteMissionDepartment = await missionService.deleteDepartmentMission(findResponsibleMissionDepartment.id);
-            if (deleteMissionDepartment) {
+        if (responsibleDepartment_id) {
+            const findResponsibleDepartment = await missionService.getDepartmentById(responsibleDepartment_id);
+            const findResponsibleMissionDepartment = await missionService.getDepartmentMissionById(getMissionById.id, true);
+            if (findResponsibleMissionDepartment) {
+                await missionService.deleteDepartmentMission(findResponsibleMissionDepartment.id);
+                await getMissionById.addDepartment(findResponsibleDepartment, { through: { isResponsible: true } });
+            } else {
                 await getMissionById.addDepartment(findResponsibleDepartment, { through: { isResponsible: true } });
             }
         }
-        const findRelatedMissionDepartment = await missionService.getDepartmentMissionById(getMissionById.id, false);
-        const findRelatedDepartment = await missionService.getDepartmentById(relatedDepartment_id);
-        if (findRelatedMissionDepartment) {
-            const deleteMissionDepartment = await missionService.deleteDepartmentMission(findRelatedMissionDepartment.id);
-            if (deleteMissionDepartment) {
+        if (relatedDepartment_id) {
+            const findRelatedDepartment = await missionService.getDepartmentById(relatedDepartment_id);
+            const findRelatedMissionDepartment = await missionService.getDepartmentMissionById(getMissionById.id, false);
+            if (findRelatedMissionDepartment) {
+                await missionService.deleteDepartmentMission(findRelatedMissionDepartment.id);
+                await getMissionById.addDepartment(findRelatedDepartment, { through: { isResponsible: false } });
+            } else {
                 await getMissionById.addDepartment(findRelatedDepartment, { through: { isResponsible: false } });
             }
         }
         return res.status(200).json({ msg: 'Success!', data: updateMission });
-        } catch (error) {
-            return res.status(400).json({ message: error });
-        }
+    } catch (error) {
+        return res.status(400).json({ message: error });
+    }
 };
